Cleanup (test moved to a correct location)
