+++ java/java-tests/testData/codeInsight/daemonCodeAnalyzer/advHighlighting8/MethodReferences.java
