+++ java/java-tests/testData/codeInsight/daemonCodeAnalyzer/advHighlighting8/UsedMethodsByMethodReferences.java
